Add testng.xml

+++ src/test/java/step_definitions/MasterHooks.java
@@ -1,7 +1,9 @@
 package step_definitions;
 
+import io.cucumber.core.api.Scenario;
 import io.cucumber.java.After;
 import io.cucumber.java.Before;
+import page_objects.BasePage;
 import utils.DriverFactory;
 
 public class MasterHooks extends DriverFactory {
@@ -12,11 +14,23 @@ public class MasterHooks extends DriverFactory {
     }
 
     @After
-    public void tearDown() {
-        if (driver != null) {
-            driver.manage().deleteAllCookies();
-            driver.quit();
+    public void tearDownAndScreenshotOnFailure(Scenario scenario) {
+        try {
+            if (driver != null && scenario.isFailed()) {
+                BasePage.captureScreenshot();
+                driver.manage().deleteAllCookies();
+                driver.quit();
+                driver = null;
+            }
+
+            if (driver != null) {
+                driver.manage().deleteAllCookies();
+                driver.quit();
+            }
+        } catch (Exception e) {
+            System.out.println("Method has failed: tearDownAndScreenshotOnFailure, Exception " + e.getMessage());
         }
+
     }
 
 }
